Moving ContrabandHitRate and RaceToggle
